performance_requirements.txt
# PERFORMANCE OPTIMIZATION DEPENDENCIES
# Add these to your requirements.txt for optimal performance

# Redis caching
django-redis==5.3.0
hiredis==2.2.3

# Background task processing
celery[redis]==5.3.0

# Development debugging (optional)
django-debug-toolbar==4.2.0
django-silk==5.0.3

# Error tracking (optional)
sentry-sdk[django]==1.38.0

# Performance monitoring (optional)
newrelic==9.2.0

# Image optimization (if not using Cloudinary)
pillow-simd==10.0.1.post1  # Faster than regular Pillow

# Compression
brotli==1.1.0

# Database optimization
psycopg2-binary==2.9.7  # Already in your requirements

# Additional useful packages
whitenoise[brotli]==6.6.0  # Upgraded version with Brotli support